+++ src/classes/managers/orders/orders.manager.ts
@@ -1,5 +1,6 @@
 import { Subject } from 'rxjs';
 import { CTraderConnection } from '@himalaya-quant/ctrader-layer';
+import { CTraderLayerEvent } from '@himalaya-quant/ctrader-layer/build/src/core/events/CTraderLayerEvent';
 
 import { ILogger } from '../../logger';
 import { BaseManager } from '../models/base.manager';
@@ -24,18 +25,22 @@ import {
     isCancelledOrder,
 } from './typeguards/positions-typeguards';
 
-import { ProtoOANewOrderReq } from './proto/messages/ProtoOANewOrderReq';
-import { ProtoOAExecutionEvent } from './proto/events/ProtoOAExecutionEvent';
-import { ProtoOAOrderErrorEvent } from './proto/events/ProtoOAOrderErrorEvent';
-import { ProtoOAGetPositionUnrealizedPnLReq } from './proto/messages/ProtoOAGetPositionUnrealizedPnLReq';
-import { ProtoOAGetPositionUnrealizedPnLRes } from './proto/messages/ProtoOAGetPositionUnrealizedPnLRes';
-import { GetPositionUnrealizedPnLError } from './errors/get-position-unrealized-pnl.error';
-import { ProtoOAPayloadType } from '../../../models/proto/payload-types/payload-types.enum';
-import { CTraderLayerEvent } from '@himalaya-quant/ctrader-layer/build/src/core/events/CTraderLayerEvent';
-import { ProtoOAClosePositionReq } from '../symbols/proto/messages/ProtoOAClosePositionReq';
 import { ClosePositionError } from './errors/close-position.error';
+import { GetOpenPositionsError } from './errors/get-open-positions.error';
+import { CancelPendingOrderError } from './errors/cancel-pending-order.error';
+import { ModifyPendingOrderError } from './errors/modify-pending-order.error';
+import { GetPositionUnrealizedPnLError } from './errors/get-position-unrealized-pnl.error';
+
+import { ProtoOANewOrderReq } from './proto/messages/ProtoOANewOrderReq';
 import { ProtoOAReconcileReq } from './proto/messages/ProtoOAReconcileReq';
 import { ProtoOAReconcileRes } from './proto/messages/ProtoOAReconcileRes';
+import { ProtoOAAmendOrderReq } from './proto/messages/ProtoOAAmendOrderReq';
+import { ProtoOAExecutionEvent } from './proto/events/ProtoOAExecutionEvent';
+import { ProtoOACancelOrderReq } from './proto/messages/ProtoOACancelOrderReq';
+import { ProtoOAOrderErrorEvent } from './proto/events/ProtoOAOrderErrorEvent';
+import { ProtoOAClosePositionReq } from '../symbols/proto/messages/ProtoOAClosePositionReq';
+import { ProtoOAGetPositionUnrealizedPnLReq } from './proto/messages/ProtoOAGetPositionUnrealizedPnLReq';
+import { ProtoOAGetPositionUnrealizedPnLRes } from './proto/messages/ProtoOAGetPositionUnrealizedPnLRes';
 
 type BaseProto = 'payloadType' | 'ctidTraderAccountId';
 
@@ -54,19 +59,64 @@ class OrdersEventsDispatcher {
 }
 
 export class OrdersManager extends BaseManager {
+    private readonly subscriptionsIds = new Set<string>();
+
     constructor(
         protected readonly credentials: ICredentials,
         protected readonly connection: CTraderConnection,
         protected readonly logger: ILogger,
     ) {
         super();
-        this.openExecutionListeners();
+        this.openEventsListeners();
+    }
+
+    /**
+     * Closes all the subscriptions, freeing memory and avoiding zombie event
+     * listeners dangling
+     */
+    dispose() {
+        this.subscriptionsIds.forEach((id) =>
+            this.connection.removeEventListener(id),
+        );
     }
 
     subscribeOrdersEvents() {
         return OrdersEventsDispatcher.subscribeEvents();
     }
 
+    /**
+     * Request for cancelling existing pending order.
+     * Allowed only if the accessToken has "trade" permissions for the trading account.
+     * @param req The details for the pending order to close
+     */
+    async cancelPendingOrder(req: Omit<ProtoOACancelOrderReq, BaseProto>) {
+        this.logCallAttempt(this.cancelPendingOrder);
+        const payload: ProtoOACancelOrderReq = {
+            ...req,
+            ctidTraderAccountId: this.credentials.ctidTraderAccountId,
+        };
+
+        try {
+            await this.connection.sendCommand(
+                ProtoOACancelOrderReq.name,
+                payload,
+            );
+        } catch (e) {
+            throw this.handleCTraderCallError(
+                e,
+                this.cancelPendingOrder,
+                new CancelPendingOrderError(e),
+            );
+        }
+
+        this.logCallAttemptSuccess(this.cancelPendingOrder);
+    }
+
+    /**
+     * Request for sending a new trading order.
+     * Allowed only if the accessToken has the "trade" permissions for the trading account.
+     * @param req The details for the new order to send
+     */
     async newOrder(req: Omit<ProtoOANewOrderReq, BaseProto>) {
         this.logCallAttempt(this.newOrder);
         const payload: ProtoOANewOrderReq = {
@@ -87,6 +137,40 @@ export class OrdersManager extends BaseManager {
         this.logCallAttemptSuccess(this.newOrder);
     }
 
+    /**
+     * Request for amending the existing pending order.
+     * Allowed only if the Access Token has "trade" permissions for the trading account.
+     * @param req The details for the pending order to modify
+     */
+    async modifyPendingOrder(req: Omit<ProtoOAAmendOrderReq, BaseProto>) {
+        this.logCallAttempt(this.modifyPendingOrder);
+        const payload: ProtoOAAmendOrderReq = {
+            ...req,
+            ctidTraderAccountId: this.credentials.ctidTraderAccountId,
+        };
+
+        try {
+            await this.connection.sendCommand(
+                ProtoOAAmendOrderReq.name,
+                payload,
+            );
+        } catch (e) {
+            throw this.handleCTraderCallError(
+                e,
+                this.modifyPendingOrder,
+                new ModifyPendingOrderError(e),
+            );
+        }
+
+        this.logCallAttemptSuccess(this.modifyPendingOrder);
+    }
+
+    /**
+     * Request for closing or partially closing of an existing position.
+     * Allowed only if the accessToken has "trade" permissions for the trading account.
+     *
+     * @param req The details for the position to close
+     */
     async closePosition(req: Omit<ProtoOAClosePositionReq, BaseProto>) {
         this.logCallAttempt(this.closePosition);
 
@@ -134,7 +218,7 @@ export class OrdersManager extends BaseManager {
             throw this.handleCTraderCallError(
                 e,
                 this.getOpenPositions,
-                new ClosePositionError(e),
+                new GetOpenPositionsError(e),
             );
         }
 
@@ -171,15 +255,19 @@ export class OrdersManager extends BaseManager {
         return result;
     }
 
-    private openExecutionListeners() {
-        this.connection.on(
-            ProtoOAOrderErrorEvent.name,
-            this.handleOrderEventError.bind(this),
+    private openEventsListeners() {
+        this.subscriptionsIds.add(
+            this.connection.on(
+                ProtoOAOrderErrorEvent.name,
+                this.handleOrderEventError.bind(this),
+            ),
         );
 
-        this.connection.on(
-            ProtoOAExecutionEvent.name,
-            this.handleOrderExecutionEvent.bind(this),
+        this.subscriptionsIds.add(
+            this.connection.on(
+                ProtoOAExecutionEvent.name,
+                this.handleOrderExecutionEvent.bind(this),
+            ),
         );
     }