+++ src/classes/managers/symbols/proto/messages/ProtoOASubscribeLiveTrendbarReq.ts
@@ -1,5 +1,5 @@
 import { BaseProto } from '../../../../../models/proto/base-proto';
-import { ProtoOATrendbarPeriod } from '../../../../../models/proto/models/ProtoOATrendbarPeriod';
+import { ProtoOATrendbarPeriod } from '../models/ProtoOATrendbarPeriod';
 
 export class ProtoOASubscribeLiveTrendbarReq extends BaseProto {
     symbolId: number;