+++ src/classes/managers/symbols/proto/messages/ProtoOASubscribeSpotsRes.ts
@@ -1,0 +1,3 @@
+import { BaseProto } from '../../../../../models/proto/base-proto';
+
+export class ProtoOASubscribeSpotsRes extends BaseProto {}