+++ src/examples/get-symbols-bars.example.ts
@@ -1,5 +1,5 @@
 import { cTraderX } from '../classes/client';
-import { ProtoOATrendbarPeriod } from '../models/proto/models/ProtoOATrendbarPeriod';
+import { ProtoOATrendbarPeriod } from '../classes/managers/symbols/proto/models/ProtoOATrendbarPeriod';
 
 // Get bars for symbol
 (async () => {